+++ src/unfold/views.py
@@ -1,8 +1,21 @@
 from typing import Any
 
+import django
+from django.contrib.admin.views.main import ERROR_FLAG, PAGE_VAR
+from django.contrib.admin.views.main import ChangeList as BaseChangeList
 from django.contrib.auth.mixins import PermissionRequiredMixin
 
-from .exceptions import UnfoldException
+from unfold.exceptions import UnfoldException
+
+
+class ChangeList(BaseChangeList):
+    def __init__(self, request, *args, **kwargs):
+        super().__init__(request, *args, **kwargs)
+
+        if django.VERSION < (5, 0):
+            self.filter_params = dict(request.GET.lists())
+            self.filter_params.pop(PAGE_VAR, None)
+            self.filter_params.pop(ERROR_FLAG, None)
 
 
 class UnfoldModelAdminViewMixin(PermissionRequiredMixin):
